requirements.txt
aiobotocore==2.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiokafka==0.12.0
aioredis==2.0.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arq==0.26.3
async-timeout==5.0.1
asyncio-mqtt==0.16.2
asyncpg==0.30.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.39.11
botocore==1.39.11
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clickhouse-connect==0.9.2
clickhouse-driver==0.2.9
ConfigArgParse==1.7.1
confluent-kafka==2.11.1
cte==0.1.0
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.3.0
Faker==37.6.0
fastapi==0.116.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
frozenlist==1.7.0
fsspec==2025.9.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kafka-python==2.2.15
locust==2.41.6
locust-cloud==1.27.5
lz4==4.4.4
MarkupSafe==3.0.3
minio==7.2.16
msgpack==1.1.2
multidict==6.6.4
numpy==2.3.3
orjson==3.11.3
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.1.0
psycopg2-binary==2.9.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
redis==5.3.1
requests==2.32.5
s3fs==2025.9.0
s3transfer==0.13.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.15
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zope.event==6.0
zope.interface==8.0.1
zstandard==0.25.0
